refactor: extract EventRepository
